New page for patients to see indepth detail about a single volunteer
This page was created so when a patient clicks on a specalist they can learn more about them before selecting to see that volunteer.

In the future there needs to be a back button added to get back to all types of volunteers.

Also General Practioners would not have the list of specialities listed on  their profile